Added all output to console needed. New method to save and print command name and description.
Added json loader to load .json config files.
Command names and descriptions are now saved on a json file, for better organization.
Command <Commands> now runs on a foreach loop, which makes it much simpler, and no need to update it.
Commands info can be both added through the command <new_c> and directly into the file.
Added "Dictionary_Path", variable that should contain the path to the file Command_Dictionary.json.
Command_Dictionary.json is mandatory to exist, so I'll upload it to the repository too. Simply download it, and copy its path to the variable.
Currently there is no method to check the existence of this file.